Listar beneficiário:   - Implementado filtro de status e de plano

+++ front-spa/src/app/pages/plano/listar-plano/listar-plano.component.ts
@@ -7,6 +7,7 @@ import { Plano } from '../../../core/models/plano'
 import { ErrorDialogService } from '../../../shared/services/error-dialog.service'
 import { ConfirmActionService } from '../../../shared/services/modal-confirmation.service'
 import { SuccessDialogService } from '../../../shared/services/success-dialog.service'
+import { BeneficiarioHttp } from '../../../core/http/beneficiario.http'
 
 @Component({
   selector: 'app-listar-plano.component',
@@ -21,6 +22,7 @@ export class ListarPlanoComponent implements OnInit {
   private router = inject(Router)
 
   private http = inject(PlanosHttp)
+  private beneficiarioHttp = inject(BeneficiarioHttp)
 
   private confirmActioncService = inject(ConfirmActionService)
   private successDialogService = inject(SuccessDialogService)
@@ -34,7 +36,7 @@ export class ListarPlanoComponent implements OnInit {
     this.router.navigate(['/plano/editar'], {queryParams: {'id': id}})
   }
 
-  onDeletePlano(id: string, index: number) { 
+  async onDeletePlano(id: string, index: number) {
     this.confirmActioncService
       .setTitle('Confirma a exclusão do plano?')
       .setOnConfirm(this.deletePlano.bind(this, id, index))
@@ -64,7 +66,21 @@ export class ListarPlanoComponent implements OnInit {
     ))
   }
 
-  private deletePlano(id: string, index: number) {
+  private async validateDelete(index: number) {
+    let result = await this.beneficiarioHttp.getByPlano(this.planos[index].Id)
+
+    if (result && result.length > 0) return false
+    return true
+  }
+
+  private async deletePlano(id: string, index: number) {
+    if (!await this.validateDelete(index)) {
+      this.errorDialogService
+          .setMessages([`Não é possível apagar este plano, pois existe pelo menos um beneficiário cadastrado nele.`])
+          .show()
+      return
+    }
+
     this.http.delete(id)
     .subscribe({
       next: (value: any) => {
